pyproject.toml
[tool.poetry]
name = "portal-sspk"
version = "0.1.0"
description = ""
authors = ["s.leontev <neonyan15@gmail.com>"]
readme = "README.md"
packages = [{ include = "portal_sspk" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
python-dotenv = "^0.21.0"
django-phonenumber-field = "^7.0.2"
phonenumbers = "^8.13.4"
requests = "^2.28.2"
django-polymorphic = "^3.1.0"
pillow = "^9.4.0"
redis = "^4.4.2"
celery = "^5.2.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
pandas = "^1.5.2"
gunicorn = "^20.1.0"
whitenoise = "^6.3.0"
xlsxwriter = "^3.0.7"
djangorestframework-simplejwt = "^5.2.2"
django-rest-polymorphic = "^0.1.10"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.11.4"
ipython = "^8.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
